Find cross-group tasks in iter_mapped_dependants
